Sorting eigenvalues. Extended potential function.
Program is sorting eigenvalues. However, they do not agree with those
given in the project information.
Potential function is now extended to include two possible potentials,
and requires input from the terminal to decide which one of them to
create. At this stage, omega_r is also taken in from terminal. It should
be considered whether to incorporate a for-loop for all possible values
or not.